# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PaySplit Backend API
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/groups:
    post:
      operationId: createGroup
      tags: [Groups]
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateGroupRequest"}
      responses:
        "201":
          description: Group created with the caller as its active Captain
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CreateGroupResponse"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
    get:
      operationId: listGroups
      tags: [Groups]
      security: [{bearerAuth: []}]
      parameters:
        - name: cursor
          in: query
          schema: {type: string}
          description: Opaque cursor from a prior response's next_cursor
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        "200":
          description: Groups where the caller is an active member, newest first
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GroupListResponse"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /api/v1/groups/{id}:
    get:
      operationId: getGroupDetail
      tags: [Groups]
      security: [{bearerAuth: []}]
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: Group, active members, balances, and the caller's role
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GroupDetailResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
    patch:
      operationId: renameGroup
      tags: [Groups]
      security: [{bearerAuth: []}]
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RenameGroupRequest"}
      responses:
        "200":
          description: Group renamed and the activity committed atomically
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GroupResponse"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/CaptainRequired"}
        "404": {$ref: "#/components/responses/NotFound"}
    delete:
      operationId: disbandGroup
      tags: [Groups]
      security: [{bearerAuth: []}]
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "204": {description: Group archived without deleting financial or audit history}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/CaptainRequired"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/GroupHasUnsettledObligations"}
...
  /api/v1/groups/join:
    post:
      operationId: joinGroup
      tags: [Groups]
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/JoinGroupRequest"}
      responses:
        "200":
          description: Joined, reactivated, or already an active member
          content:
            application/json:
              schema: {$ref: "#/components/schemas/JoinGroupResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/InviteNotFound"}
        "409": {$ref: "#/components/responses/GroupMemberLimitReached"}
        "429": {$ref: "#/components/responses/RateLimited"}
  /api/v1/groups/{id}/members/{memberId}:
    delete:
      operationId: leaveOrRemoveMember
      tags: [Groups]
      security: [{bearerAuth: []}]
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - name: memberId
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "204": {description: Membership marked inactive, idempotent on retry}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409":
          description: Target has unsettled debts (GROUP_MEMBER_HAS_OPEN_DEBTS), or is the active Captain and must transfer the role first (CAPTAIN_TRANSFER_REQUIRED)
          content:
            application/json:
              schema:
                oneOf:
                  - {$ref: "#/components/schemas/OpenDebtsError"}
                  - {$ref: "#/components/schemas/Error"}
  /api/v1/groups/{id}/members/{memberId}/role:
    put:
      operationId: transferCaptain
      tags: [Groups]
      security: [{bearerAuth: []}]
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - name: memberId
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/TransferRoleRequest"}
      responses:
        "200":
          description: Captain role transferred to the target member
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CaptainTransferResponse"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/CaptainRequired"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409":
          description: The group lock for this transfer could not be acquired immediately
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
...
  responses:
    BadRequest:
      description: Validation or invalid token error
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
    Unauthorized:
      description: Authentication or session error
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
    Forbidden:
      description: Account is not eligible
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
...
    NotFound:
      description: Resource missing, or the caller is not authorized to see it
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
...
    CaptainRequired:
      description: Caller is not the group's active Captain
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
    InviteNotFound:
      description: Invite is malformed, unknown, expired, revoked, or exhausted; every unusable state returns INVITE_NOT_FOUND
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
    GroupHasUnsettledObligations:
      description: Group has bills outside finalized or voided, or debts outside settled or voided
      content: {application/json: {schema: {$ref: "#/components/schemas/GroupUnsettledObligationsError"}}}
    GroupMemberLimitReached:
      description: Group already has 50 active members
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
    RateLimited:
      description: Rolling limit exceeded
      headers:
        Retry-After: {schema: {type: integer}, description: Seconds until retry}
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
            fields: {type: object, additionalProperties: {type: string}}
    OpenDebtsError:
      description: Error variant for GROUP_MEMBER_HAS_OPEN_DEBTS, carrying the target's outstanding amounts
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, fields]
          properties:
            code: {type: string, enum: [GROUP_MEMBER_HAS_OPEN_DEBTS]}
            message: {type: string}
            fields:
              type: object
              required: [payable_amount, receivable_amount]
              properties:
                payable_amount: {type: string, description: Decimal VND string the target still owes others}
                receivable_amount: {type: string, description: Decimal VND string others still owe the target}
    GroupUnsettledObligationsError:
      description: Error variant for GROUP_HAS_UNSETTLED_OBLIGATIONS
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, fields]
          properties:
            code: {type: string, enum: [GROUP_HAS_UNSETTLED_OBLIGATIONS]}
            message: {type: string}
            fields:
              type: object
              required: [draft_or_reviewed_bill_count, open_debt_count]
              properties:
                draft_or_reviewed_bill_count: {type: string, pattern: '^[0-9]+$'}
                open_debt_count: {type: string, pattern: '^[0-9]+$'}
    CreateGroupRequest:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name: {type: string, minLength: 1, maxLength: 100, description: Trimmed to 1-100 Unicode code points}
        currency: {type: string, enum: [VND], description: Optional; only VND is accepted in v1}
    Group:
      type: object
      required: [id, name, currency, created_at]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        currency: {type: string, enum: [VND]}
        created_at: {type: string, format: date-time}
    RenameGroupRequest:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name: {type: string, minLength: 1, maxLength: 100, description: Trimmed to 1-100 Unicode code points}
    GroupResponse:
      type: object
      required: [group]
      properties:
        group: {$ref: "#/components/schemas/Group"}
    Membership:
      type: object
      required: [id, group_id, user_id, role, status, joined_at, left_at]
      properties:
        id: {type: string, format: uuid}
        group_id: {type: string, format: uuid}
        user_id: {type: string, format: uuid}
        role: {type: string, enum: [captain, member]}
        status: {type: string, enum: [active, inactive]}
        joined_at: {type: string, format: date-time}
        left_at: {type: string, format: date-time, nullable: true}
    Member:
      type: object
      required: [membership_id, user_id, display_name, avatar_url, role, joined_at]
      properties:
        membership_id: {type: string, format: uuid}
        user_id: {type: string, format: uuid}
        display_name: {type: string}
        avatar_url: {type: string, format: uri, nullable: true}
        role: {type: string, enum: [captain, member]}
        joined_at: {type: string, format: date-time}
    Balance:
      type: object
      required: [member_id, net_balance]
      properties:
        member_id: {type: string, format: uuid}
        net_balance: {type: string, description: Decimal VND string; positive means the member is owed}
    GroupListItem:
      type: object
      required: [group, caller_membership_id, caller_role, active_member_count]
      properties:
        group: {$ref: "#/components/schemas/Group"}
        caller_membership_id: {type: string, format: uuid}
        caller_role: {type: string, enum: [captain, member]}
        active_member_count: {type: integer}
    CreateGroupResponse:
      type: object
      required: [group, membership]
      properties:
        group: {$ref: "#/components/schemas/Group"}
        membership: {$ref: "#/components/schemas/Membership"}
    GroupListResponse:
      type: object
      required: [groups, next_cursor]
      properties:
        groups: {type: array, items: {$ref: "#/components/schemas/GroupListItem"}}
        next_cursor: {type: string, nullable: true}
    GroupDetailResponse:
      type: object
      required: [group, members, balances, caller_role]
      properties:
        group: {$ref: "#/components/schemas/Group"}
        members: {type: array, items: {$ref: "#/components/schemas/Member"}}
        balances: {type: array, items: {$ref: "#/components/schemas/Balance"}}
        caller_role: {type: string, enum: [captain, member]}
...
    JoinGroupRequest:
      type: object
      required: [code]
      additionalProperties: false
      properties:
        code: {type: string, minLength: 8, maxLength: 8, pattern: '^[A-Za-z0-9]{8}$', description: Malformed values still receive INVITE_NOT_FOUND}
    JoinResult:
      type: object
      required: [group_id, membership_id, role, status, result]
      properties:
        group_id: {type: string, format: uuid}
        membership_id: {type: string, format: uuid}
        role: {type: string, enum: [captain, member]}
        status: {type: string, enum: [active, inactive]}
        result: {type: string, enum: [joined, reactivated, already_active]}
    JoinGroupResponse:
      type: object
      required: [join]
      properties:
        join: {$ref: "#/components/schemas/JoinResult"}
    TransferRoleRequest:
      type: object
      required: [role]
      additionalProperties: false
      properties:
        role: {type: string, enum: [captain]}
    CaptainTransfer:
      type: object
      required: [previous_captain_member_id, current_captain_member_id]
      properties:
        previous_captain_member_id: {type: string, format: uuid}
        current_captain_member_id: {type: string, format: uuid}
    CaptainTransferResponse:
      type: object
      required: [transfer]
      properties:
        transfer: {$ref: "#/components/schemas/CaptainTransfer"}
...